Geometry calculations now use floats Added solid type prototype for 3D rendering

+++ include/Graphics/geometry.hpp
@@ -3,19 +3,20 @@
 #include <iostream>
 #include <vector>
 #include <memory>
-#include <glm/glm.hpp>
+#include <glm/ext/vector_float3.hpp>
+#include <glm/geometric.hpp>
 
 class Point {
     public:
         std::vector<Point> vertices;
-        glm::dvec3 pos;
+        glm::vec3 pos;
         Point();
-        Point(glm::dvec3 pos);
+        Point(glm::vec3 pos);
         virtual unsigned int dim() const {return 0;}
         virtual bool isSpace() const {return true;}
-        virtual double dist(const Point &obj) const;
+        virtual float dist(const Point &obj) const;
         virtual std::unique_ptr<Point> intersect (const Point &obj) const;
-        glm::dvec3 direction(const Point &obj) const;
+        glm::vec3 direction(const Point &obj) const;
         bool contains(const Point &obj) const;
         bool equals(const Point &obj) const;
 };
@@ -26,13 +27,13 @@ class Line: public Point {
         Line(Point p1, Point p2);
         Line(std::vector<Point> vert);
         unsigned int dim() const override {return 1;}
-        double dist(const Point &obj) const override;
-        virtual double dist(const Line &obj) const;
+        float dist(const Point &obj) const override;
+        virtual float dist(const Line &obj) const;
         std::unique_ptr<Point> intersect(const Point &obj) const override;
         virtual std::unique_ptr<Point> intersect(const Line &obj) const;
-        glm::dvec3 dirVec() const;
+        glm::vec3 dirVec() const;
         std::unique_ptr<Point> project(const Point &obj) const;
-        double angle(const Line &lobj, glm::dvec3* axisptr);
+        float angle(const Line &lobj, glm::vec3* axisptr);
 };
 
 class LinSeg: public Line {
@@ -41,13 +42,13 @@ class LinSeg: public Line {
         LinSeg(Point p1, Point p2);
         LinSeg(std::vector<Point> vert);
         bool isSpace() const override {return false;}
-        double dist(const Point &obj) const override;
-        double dist(const Line &obj) const override;
-        double dist(const LinSeg &obj) const;
+        float dist(const Point &obj) const override;
+        float dist(const Line &obj) const override;
+        float dist(const LinSeg &obj) const;
         std::unique_ptr<Point> intersect(const Point &obj) const override;
         std::unique_ptr<Point> intersect(const Line &obj) const override;
         std::unique_ptr<Point> intersect(const LinSeg &obj) const;
-        double length() const;
+        float length() const;
 };
 
 class Plane: public Point {
@@ -56,7 +57,7 @@ class Plane: public Point {
         Plane(Point p1, Point p2, Point p3);
         Plane(std::vector<Point> vert);
         unsigned int dim() const override {return 2;}
-        glm::dvec3 normVec() const;
+        glm::vec3 normVec() const;
         std::unique_ptr<Point> project(const Point &obj) const;
         template<typename T>
         std::unique_ptr<T> project(const T &obj) const {
@@ -64,11 +65,11 @@ class Plane: public Point {
             for(Point p: obj.vertices) v.push_back(*project(p));
             return std::unique_ptr<T>(new T(v));
         }
-        double sign_dist(const Point &obj) const;
-        double dist(const Point &obj) const override;
-        virtual double dist(const Line &obj) const;
-        virtual double dist(const LinSeg &obj) const;
-        virtual double dist(const Plane &obj) const;
+        float sign_dist(const Point &obj) const;
+        float dist(const Point &obj) const override;
+        virtual float dist(const Line &obj) const;
+        virtual float dist(const LinSeg &obj) const;
+        virtual float dist(const Plane &obj) const;
         std::unique_ptr<Point> intersect(const Point &obj) const override;
         virtual std::unique_ptr<Point> intersect(const Line &obj) const;
         virtual std::unique_ptr<Point> intersect(const LinSeg &obj) const;
@@ -83,17 +84,17 @@ class Polygon: public Plane {
         Polygon(Point p1, Point p2, Point p3, Points... args): Polygon(std::vector<Point>{p1, p2, p3, args...}){};
         Polygon(std::vector<Point> vert);
         bool isSpace() const override {return false;}
-        double dist(const Point &obj) const override;
-        double dist(const Line &obj) const override;
-        double dist(const LinSeg &obj) const override;
-        double dist(const Plane &obj) const override;
-        double dist(const Polygon &obj) const;
+        float dist(const Point &obj) const override;
+        float dist(const Line &obj) const override;
+        float dist(const LinSeg &obj) const override;
+        float dist(const Plane &obj) const override;
+        float dist(const Polygon &obj) const;
         std::unique_ptr<Point> intersect(const Point &obj) const override;
         std::unique_ptr<Point> intersect(const Line &obj) const override;
         std::unique_ptr<Point> intersect(const LinSeg &obj) const override;
         std::unique_ptr<Point> intersect(const Plane &obj) const override;
         std::unique_ptr<Point> intersect(const Polygon &obj) const;
-        double area() const;
+        float area() const;
 };
 
 class Polyhedron: public Point {
@@ -102,23 +103,23 @@ class Polyhedron: public Point {
         std::vector<Polygon> faces;
         Polyhedron(){};
         template <typename... Points>
-        Polyhedron(Point p1, Point p2, Point p3, Point p4, Points... args): Polyhedron(std::vector<Point>{p1, p2, p3, p4, args...}) {};
+        Polyhedron(Point p1, Point p2, Point p3, Point p4, Points... args): Polyhedron(std::vector<Point>{p1, p2, p3, p4, args...}){};
         Polyhedron(std::vector<Point> vert);
         unsigned int dim() const override {return 3;}
         bool isSpace() const override {return false;}
-        double dist(const Point &obj) const override;
-        double dist(const Line &obj) const;
-        double dist(const LinSeg &obj) const;
-        double dist(const Plane &obj) const;
-        double dist(const Polygon &obj) const;
-        double dist(const Polyhedron &obj) const;
+        float dist(const Point &obj) const override;
+        float dist(const Line &obj) const;
+        float dist(const LinSeg &obj) const;
+        float dist(const Plane &obj) const;
+        float dist(const Polygon &obj) const;
+        float dist(const Polyhedron &obj) const;
         std::unique_ptr<Point> intersect(const Point &obj) const override;
         // std::unique_ptr<Point> intersect(const Line &obj) const;
         // std::unique_ptr<Point> intersect(const LinSeg &obj) const;
         // std::unique_ptr<Point> intersect(const Plane &obj) const;
         // std::unique_ptr<Point> intersect(const Polygon &obj) const;
         // std::unique_ptr<Point> intersect(const Polyhedron &obj) const;
-        double volume() const;
+        float volume() const;
 };
 
 static std::ostream& operator<<(std::ostream &strm, const Point &p){
